+++ api/v1alpha1/token_types.go
@@ -27,7 +27,7 @@ type TokenSpec struct {
 
 	// Slurm `auth/jwt` JWT HS256 key authentication.
 	// +required
-	JwtHs256KeyRef corev1.SecretKeySelector `json:"jwtHs256KeyRef,omitzero"`
+	JwtHs256KeyRef JwtSecretKeySelector `json:"jwtHs256KeyRef,omitzero"`
 
 	// The username whom the token is created for.
 	// +required