+++ internal/builder/token_secret.go
@@ -16,7 +16,8 @@ import (
 func (b *Builder) BuildTokenSecret(token *slinkyv1alpha1.Token) (*corev1.Secret, error) {
 	ctx := context.TODO()
 
-	signingKey, err := b.refResolver.GetSecretKeyRef(ctx, token.JwtHs256Ref(), token.Namespace)
+	jwtHs256Ref := token.JwtHs256Ref()
+	signingKey, err := b.refResolver.GetSecretKeyRef(ctx, &jwtHs256Ref.SecretKeySelector, jwtHs256Ref.Namespace)
 	if err != nil {
 		return nil, err
 	}