+++ internal/controller/token/token_sync.go
@@ -62,7 +62,8 @@ func (r *TokenReconciler) Sync(ctx context.Context, req reconcile.Request) error
 				if err != nil {
 					return err
 				}
-				signingKey, err := r.refResolver.GetSecretKeyRef(ctx, token.JwtHs256Ref(), token.Namespace)
+				jwtHs256Ref := token.JwtHs256Ref()
+				signingKey, err := r.refResolver.GetSecretKeyRef(ctx, &jwtHs256Ref.SecretKeySelector, jwtHs256Ref.Namespace)
 				if err != nil {
 					return err
 				}